requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
newspaper3k==0.2.8
beautifulsoup4==4.12.3
requests==2.31.0
python-dotenv==1.0.1
openai==1.12.0
pydantic==2.6.1 